+++ solagent-adapters/rig/solayer/Cargo.toml
@@ -1,17 +1,17 @@
 [package]
 name = "solagent-rig-solayer"
-version = "0.1.1"
+version = "0.1.2"
 edition = "2021"
 authors = ["zTgx <beautifularea@gmail.com>"]
 repository = "https://github.com/zTgx/solagent.rs"
 keywords = ["solagent", "rig", "solayer"]
 license = "Apache-2.0"
-description = "solagent.rs rig solayer"
+description = "solagent rig solayer"
 
 [dependencies]
-solagent-core = "0.1.3"
+solagent-core = "0.1.6"
+solagent-plugin-solayer = "0.1.2"
 solagent-parameters = "0.1.0"
-solagent-plugin-solayer = "0.1.1"
 serde = { version = "1.0", features = ["derive"] }
 thiserror = "2.0.9"
 serde_json = "1.0"
